% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAFE.R
\name{SAFE}
\alias{SAFE}
\title{Single-cell Aggregated clustering From Ensemble (SAFE)}
\usage{
SAFE(cluster_results, program.dir = ".", k_min = 2, k_max = NULL,
  MCLA = TRUE, HGPA = FALSE, CSPA = FALSE, cspc_cell_max = NULL,
  SEED = 1)
}
\arguments{
\item{cluster_results}{a J*N matrix with J individual clustering methods and N cells}

\item{program.dir}{defines the directory of two programs shmetis and gpmetis, which are required by HGPA algorithm and MCLA and CSPA algorithms, respectively.
Default is the current directory (".").}

\item{k_min}{defines the minimum number of clusters used for ensembel clustering. Default is 2.}

\item{k_max}{defines the maximum number of clusters used for ensembel clustering. Default is
the maximum cluster number estimated by all the single solutions.}

\item{MCLA}{a boolean parameter that defines whether to use MCLA algorithm for ensemble clustering.
Default is "TRUE".}

\item{HGPA}{a boolean parameter that defines whether to use HGPA algorithm for ensemble clustering.
Default is "FALSE".}

\item{CSPA}{a boolean parameter that defines whether to use CSPA algorithm for ensemble clustering.
Default is "FALSE".}

\item{cspc_cell_max}{defines the maximum number of cells above which CSPA is not run, when \code{CSPA = TRUE}.}

\item{SEED}{sets the seed of the random number generator. Setting the seed to a fixed value can produce
reproducible clustering results for MCLA and CSPA algorithms.}
}
\value{
SAFE returns a list object containing:

\itemize{
    \item optimal_clustering: optimal ensemble clustering result determined by ANMI
    \item optimal_k: optimal cluster number
    \item HGPA: optimal cluster result using HGPA algorithm
    \item HGPA_optimal_k: optimal cluster number estimated by HGPA algorithm
    \item HGPA_ANMI: Average normalized mutual information (ANMI) for the optimal cluster result using HGPA algorithm
    \item MCLA: optimal cluster result using MCLA algorithm
    \item MCLA_optimal_k: optimal cluster number estimated by MCLA algorithm
    \item MCLA_ANMI: ANMI for the optimal cluster result using MCLA algorithm
    \item CSPA: optimal cluster result using CSPA algorithm
    \item CSPA_optimal_k: optimal cluster number estimated by CSPA algorithm
    \item CSPA_ANMI: ANMI for the optimal cluster result using CSPA algorithm
    \item Summary: a summary of the ensemble clustering result
}
}
\description{
SAFE (Single-cell Aggregated clustering From Ensemble): Cluster ensemble for single-cell RNA-seq data
}
\examples{
# Load the example data data_SAFE
data("data_SAFE")
working.dir = "~/Documents/single_cell_clustering"

# Zheng dataset
# Run individual_clustering
cluster.result <- individual_clustering(inputTags=data_SAFE$Zheng.expr, SEED=123)

# Cluster ensemble using SAFE clustering:
cluster.ensemble <- SAFE(cluster_results=cluster.result, program.dir = working.dir, SEED=123)

# Biase dataset
# Run individual_clustering
cluster.result <- individual_clustering(inputTags = data_SAFE$Biase.expr, datatype = "FPKM", seurat_min_cell = 200, resolution_min = 1.2, tsne_min_cells = 200, tsne_min_perplexity = 10, SEED=123)

# Cluster ensemble using SAFE clustering:
cluster.ensemble <- SAFE(cluster_results=cluster.result, program.dir = working.dir, SEED=123)
}
\references{
Yuchen Yang, Ruth Huh, Houston Culpepper, Yuan Lin, Michael Love, Yun Li. SAFE (Single-cell Aggregated clustering From Ensemble): Cluster ensemble for single-cell RNA-seq data. 2017
}
\author{
Yuchen Yang <yangyuchensysu@gmail.com>, Ruth Huh <rhuh@live.unc.edu>, Yun Li <yunli@med.unc.edu>
}
