% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_clustering.R
\name{individual_clustering}
\alias{individual_clustering}
\title{Single-cell Aggregated clustering From Ensemble (SAFE)}
\usage{
individual_clustering(inputTags, mt_filter = TRUE, mt.cutoff = 0.1, 
  SC3 = TRUE, gene_filter = FALSE, svm_num_cells = 5000, CIDR = TRUE, nPC.cidr = NULL, 
  Seurat = TRUE, nGene_filter = TRUE, low.genes = 200, high.genes = 8000, nPC.seurat = NULL, resolution = 0.7,
  tSNE = TRUE, saver = FALSE, dimensions = 3, perplexity = 30, tsne_min_cells = 200, tsne_min_perplexity = 10,
  var_genes = NULL, SEED = 1)
}
\arguments{
\item{inputTags}{a G*N matrix with G genes and N cells.}

\item{mt_filter}{is a boolean variable that defines whether to filter outlier cells according to mitochondrial gene percentage.
Default is "TRUE".}

\item{mt.pattern}{defines the pattern of mitochondrial gene names in the data, for example, \code{mt.pattern = "^MT-"} for human and \code{mt.pattern = "^mt-"} for mouse.
Default is \code{mt.pattern = "^MT-"}.}

\item{mt.cutoff}{defines a high cutoff of mitochondrial percentage (Default is 10%) that cells having higher percentage of mitochondrial gene are filtered out, when \code{mt_filter = TRUE}.}

\item{SC3}{a boolean variable that defines whether to cluster cells using SC3 method.
Default is "TRUE".}

\item{gene_filter}{a boolean variable defines whether to perform gene filtering before SC3 clustering, when \code{SC3 = TRUE}.}

\item{svm_num_cells, }{if \code{SC3 = TRUE}, then defines the mimimum number of cells above which SVM will be run.}

\item{CIDR}{a boolean parameter that defines whether to cluster cells using CIDR method.
Default is "TRUE".}

\item{nPC.cidr}{defines the number of principal coordinates used in CIDR clustering, when \code{CIDR = TRUE}.
Default value is esimated by \code{nPC} of \code{CIDR}.}

\item{Seurat}{is a boolean variable that defines whether to cluster cells using Seurat method.
Default is "TRUE".}

\item{nGene_filter}{is a boolean variable that defines whether to filter outlier cells according to unique gene count before Seurat clustering.
Default is "TRUE".}

\item{low.genes}{defines a low cutoff of unique gene counts (Default is 200) that cells having less than 200 genes are filtered out, when \code{nGene_filter = TRUE}.}

\item{high.genes}{defines a high cutoff of unique gene counts (Default is 8000) that cells having more than 8000 genes are filtered out, when \code{nGene_filter = TRUE}.}

\item{nPC.seurat}{defines the number of principal components used in Seurat clustering, when \code{Seurat = TRUE}.
Default is \code{nPC.seurat = nPC.cidr}.}

\item{resolution}{defines the value of resolution used in Seurat clustering, when \code{Seurat = TRUE}.}

\item{tSNE}{is a boolean variable that defines whether to cluster cells using t-SNE method.
Default is "TRUE".}

\item{saver}{is a boolean variable that defines whether to revise the gene expression profile in noisy and sparse single-cell RNA-seq data for downstream tSNE analysis using SAVER method.
Default is "FALSE".}

\item{dimensions}{sets the number of dimensions wanted to be retained in t-SNE step. Default is 3.}

\item{perplexity}{sets the perplexity parameter for t-SNE dimension reduction. Default is 30 when number of cells \code{>=200}.}

\item{tsne_min_cells}{defines the number of cells in input dataset below which
\code{tsne_min_perplexity=10} would be employed for t-SNE step. Default is 200.}

\item{tsne_min_perplexity}{sets the perplexity parameter of t-SNE step for small datasets (number of cells \code{<200}).}

\item{var_genes}{defines the number of variable genes used by t-SNE analysis, when \code{tSNE = TRUE}.}

\item{SEED}{sets the seed of the random number generator. Setting the seed to a fixed value can
produce reproducible clustering results.}
}
\value{
a matrix of indiviudal clustering results is output, where each row represents the cluster results of each method.
}
\description{
This function performs single-cell clustering using four state-of-the-art methods,
SC3, CIDR, Seurat and tSNE+kmeans.
}
\examples{
# Load the example data data_SAFE
data("data_SAFE")

# Zheng dataset
# Run individual_clustering
cluster.result <- individual_clustering(inputTags=data_SAFE$Zheng.expr, SEED=123)

\references{
Yuchen Yang, Ruth Huh, Houston Culpepper, Yuan Lin, Michael Love, Yun Li. SAFE (Single-cell Aggregated clustering From Ensemble): Cluster ensemble for single-cell RNA-seq data. 2017
}
\author{
Yuchen Yang <yangyuchensysu@gmail.com>, Ruth Huh <rhuh@live.unc.edu>, Yun Li <yunli@med.unc.edu>
}
